data: Move ignore file editing
Block and unblock workflows edit repository ignore files when users hide or
re-include paths. The Git utility module provides process execution,
repository discovery, index operations, and object access.

Ignore-file editing sits inside utils.git beside command execution helpers.
That forces workflow code to depend on a broad Git utility surface for
state-file edits that belong with repository data management.

This commit begins moving ignore-file state handling out of Git command
utilities by adding data.ignore_files, removing those helpers from
utils.git, and pointing block-file and unblock-file at the data module.

Subsequent commits will move the ignore-file tests to the data layer and
add a boundary guard for the old utility path.
